python/todo_api/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
psycopg2==2.9.10
sqlparse==0.5.3
tzdata==2025.2
django-cors-headers==4.7.0